+++ src/main/java/dev/gacbl/logicore/core/ModTags.java
@@ -13,6 +13,7 @@ public class ModTags {
         public static final TagKey<Block> VALID_DATACENTER_INNER_BLOCK = createTag("valid_datacenter_inner_block");
         public static final TagKey<Block> IS_ENERGY_GENERATOR = createTag("is_energy_generator");
         public static final TagKey<Block> IS_ENERGY_CABLE = createTag("is_energy_cable");
+        public static final TagKey<Block> RECYCLER_BLACKLIST = createTag("recycler_blacklist");
 
         private static TagKey<Block> createTag(String name) {
             return BlockTags.create(ResourceLocation.fromNamespaceAndPath(LogiCore.MOD_ID, name));